requirements/test.txt
pytest>=3.0.3
mock==2.0.0
requests_mock==1.1.0
